requirements.txt
Cython==0.29.32
geohash-hilbert==1.4.0
geopip==1.1
joblib==1.1.0
numpy==1.23.2
pandas==1.4.3
pyarrow==9.0.0
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
six==1.16.0